requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
Flask-Cors==3.0.10
psycopg2-binary==2.9.6
python-dotenv==1.0.0
gunicorn==20.1.0
sqlalchemy==1.4.46